accounts/serializers.py
from rest_framework import serializers
from .models import User

class UserSignupSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ( 'username', 'full_name', 'password')
