requirements.txt
altgraph==0.17
asgiref==3.2.7
astroid==2.3.3
boto3==1.13.12
botocore==1.16.12
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
fbs==0.8.6
future==0.18.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.10.0
joblib==0.14.1
lazy-object-proxy==1.4.3
llvmlite==0.31.0
lxml==4.5.0
macholib==1.14
mccabe==0.6.1
numpy==1.18.1
packaging==20.3
pefile==2019.4.18
Pillow==7.0.0
PyInstaller==3.4
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-docx==0.8.10
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
sip==5.2.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.11.2
django-heroku
